Use proxies to register types and added unregister method.
Added a macro and a proxy class to ease the process by which types can be registered with the factory. Added an unregisterType method which will be called when the proxy is deleted (e.g., when a type library is unloaded). Changed the factory internal types to be pointers to eliminate redundant copy-by-value operations and fixed creator deletion issue. Added abstract destructors to prevent memory leaks.