Add linked binaries of g2o and g2o_viewer
This commit provides linked binaries of g2o and g2o_viewer which should
allow to have some operational command line and GUI if g2o itself is
statically linked (often the case on MSVC).

g2o_linked and g2o_viewer_linked are linked against a sub-set of solvers
and types. The selection is more or less arbitrary right now and may
change in the future.