fix bug in SE3Quat::log
The value of  d  is not checked against values close to  -1. Hence, an rotation
with angle close to  pi  fails the check for performing the taylor approximation
based computation in the default case and enters the else case. However, here
one divides by  sqrt(1-d*d)  resulting in  inf  and  nan  for rotation and
translation elements of the resulting Vector6d. Checking the absolute value of
d  in the first place solves this issue.