Update README.md
There is an error in the README. Basically Visual Studio 14 2017 should not exist. To the best of my knowledge and according to [Wikipedia|https://en.wikipedia.org/wiki/Microsoft_Visual_Studio] there never was a Visual Studio 14 2017. If there actually was such a version, then it is outdated.
Therefore changed the version to the most plausible alternative. Also added a little note that one can and should change the version number.